Added environment vairables for user control of output format
Added the environment variables `fieldFormatString` and
`historyFormatString` which allow the user to specify the output format
of field and hisotry output files, respectively.

Also fixed a minor bug where the value of the field variable `YIELD` was
printed with the format `%.3f` for BS 7608 analysis, when it should
always be printed as an integer (`%.0f`).